+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Removed `params` data from `State` objects and snapshot.
 - Removed public `Storage` type export.
+- Removed `Supervisor`.`getStorage` from public API.
 
 ## [0.4.0] - 2019-10-10
 ### Added